backend/requirements.txt
alembic==1.13.3
email-validator==2.2.0
fastapi==0.115.5
httpx==0.27.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
psycopg[binary]==3.2.3
pyjwt==2.10.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose==3.3.0
sendgrid==6.11.0
sqlmodel==0.0.22
structlog==24.4.0
uvicorn[standard]==0.32.0
stripe==11.2.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-core==0.3.0
openai==1.54.0
python-dateutil==2.9.0
python-multipart==0.0.12